callcentermanager/requirements.txt
Django==4.0.5
djangorestframework==3.14.0
PyMySQL==1.0.2
pyyaml
uritemplate
corsheaders